[fix] [ugly] use only one connection to db
I downloaded your library and I liked it a lot. Sadly it was making 1 db connection for each model call. So my web host kept returning "too many connections to db" error. So I made this (yes, it's ugly) fix to make it work with only one connection. I'd love to update it in my project if you have a prettier solution.